Simplify the job pool to not reuse goroutines
Simplifies the job pool so that instead of reusing goroutines between
every round, instead we just spin up new ones every time. This has the
effect of reducing the number of synchronization primitives that we
need, and will hopefully squash a deadlock bug that I've been running
into in CI, can't reproduce locally, and can't quite figure why is
happening in a conceptual sense.

Goroutine creation is very fast, so this will have a negligible
performance impact on Modulir build loops.